refactor(host): 拆分 GitHub 宿主适配层
