+++ .loom/bin/loom_check.py
@@ -171,6 +171,7 @@ CONSUMER_REQUIRED_PATHS = (
     ".loom/bin/fact_chain_support.py",
     ".loom/bin/governance_surface.py",
     ".loom/bin/github_admission.py",
+    ".loom/bin/github_host.py",
     ".loom/bin/loom_check.py",
     ".loom/companion/manifest.json",
     ".loom/companion/repo-interface.json",
@@ -400,6 +401,7 @@ DEMO_ASSETS = (
     "examples/new-project/.loom/bin/loom_init.py",
     "examples/new-project/.loom/bin/fact_chain_support.py",
     "examples/new-project/.loom/bin/github_admission.py",
+    "examples/new-project/.loom/bin/github_host.py",
     "examples/new-project/.loom/bin/runtime_paths.py",
     "examples/new-project/.loom/bin/runtime_state.py",
     "examples/new-project/.loom/bin/loom_flow.py",
