+++ .loom/bin/loom_init.py
@@ -38,6 +38,7 @@ STORY_CARRIERS_RUNTIME_SOURCE = "skills/shared/scripts/loom_story_carriers.py"
 FACT_CHAIN_RUNTIME_SOURCE = "skills/shared/scripts/fact_chain_support.py"
 GOVERNANCE_RUNTIME_SOURCE = "skills/shared/scripts/governance_surface.py"
 ADMISSION_RUNTIME_SOURCE = "skills/shared/scripts/github_admission.py"
+GITHUB_HOST_RUNTIME_SOURCE = "skills/shared/scripts/github_host.py"
 TOOL_VERSION = "1.3.0"
 CONTRACT_VERSION = "1.3.0"
 WORK_ITEM_ID = "INIT-0001"
@@ -145,6 +146,7 @@ RUNTIME_ARTIFACT_SOURCES = {
     ".loom/bin/fact_chain_support.py": FACT_CHAIN_RUNTIME_SOURCE,
     ".loom/bin/governance_surface.py": GOVERNANCE_RUNTIME_SOURCE,
     ".loom/bin/github_admission.py": ADMISSION_RUNTIME_SOURCE,
+    ".loom/bin/github_host.py": GITHUB_HOST_RUNTIME_SOURCE,
     ".loom/bin/loom_flow.py": FLOW_RUNTIME_SOURCE,
     ".loom/bin/loom_status.py": STATUS_RUNTIME_SOURCE,
     ".loom/bin/runtime_paths.py": "skills/shared/scripts/runtime_paths.py",
@@ -1607,6 +1609,7 @@ def profile_common_artifacts(global_cli_metadata_only: bool = False) -> list[dic
             runtime_artifact(".loom/bin/fact_chain_support.py", "loom-tool-support", FACT_CHAIN_RUNTIME_SOURCE),
             runtime_artifact(".loom/bin/governance_surface.py", "loom-tool-support", GOVERNANCE_RUNTIME_SOURCE),
             runtime_artifact(".loom/bin/github_admission.py", "loom-tool-support", ADMISSION_RUNTIME_SOURCE),
+            runtime_artifact(".loom/bin/github_host.py", "loom-tool-support", GITHUB_HOST_RUNTIME_SOURCE),
             runtime_artifact(".loom/bin/loom_flow.py", "loom-tool", FLOW_RUNTIME_SOURCE),
             runtime_artifact(".loom/bin/loom_status.py", "loom-tool", STATUS_RUNTIME_SOURCE),
             runtime_artifact(".loom/bin/runtime_paths.py", "loom-tool-support", "skills/shared/scripts/runtime_paths.py"),
@@ -1685,6 +1688,7 @@ def attach_only_artifact_paths(target_root: Path, install_pr_template: bool, glo
                 ".loom/bin/fact_chain_support.py",
                 ".loom/bin/governance_surface.py",
                 ".loom/bin/github_admission.py",
+                ".loom/bin/github_host.py",
                 ".loom/bin/loom_flow.py",
                 ".loom/bin/runtime_paths.py",
                 ".loom/bin/runtime_state.py",
@@ -1728,7 +1732,7 @@ def initial_work_items(
     artifacts = [
         path
         for path in artifact_paths(initial_artifacts(target_root, install_pr_template, adoption_path, profile, global_cli_metadata_only))
-        if path != ".loom/bin/github_admission.py"
+        if path not in {".loom/bin/github_admission.py", ".loom/bin/github_host.py"}
     ]
     if profile == "light-governance":
         return [
@@ -2960,6 +2964,7 @@ def runtime_artifact(path: str, kind: str, source: str) -> dict[str, str]:
         ".loom/bin/fact_chain_support.py": Path(__file__).with_name("fact_chain_support.py"),
         ".loom/bin/governance_surface.py": Path(__file__).with_name("governance_surface.py"),
         ".loom/bin/github_admission.py": Path(__file__).with_name("github_admission.py"),
+        ".loom/bin/github_host.py": Path(__file__).with_name("github_host.py"),
         ".loom/bin/loom_flow.py": Path(__file__).with_name("loom_flow.py"),
         ".loom/bin/loom_status.py": Path(__file__).with_name("loom_status.py"),
         ".loom/bin/runtime_paths.py": Path(__file__).with_name("runtime_paths.py"),
@@ -3088,6 +3093,7 @@ def scaffold_target(
             (Path(__file__).with_name("fact_chain_support.py"), target_root / ".loom/bin/fact_chain_support.py"),
             (Path(__file__).with_name("governance_surface.py"), target_root / ".loom/bin/governance_surface.py"),
             (Path(__file__).with_name("github_admission.py"), target_root / ".loom/bin/github_admission.py"),
+            (Path(__file__).with_name("github_host.py"), target_root / ".loom/bin/github_host.py"),
             (Path(__file__).with_name("loom_flow.py"), target_root / ".loom/bin/loom_flow.py"),
             (Path(__file__).with_name("loom_status.py"), target_root / ".loom/bin/loom_status.py"),
             (Path(__file__).with_name("runtime_paths.py"), target_root / ".loom/bin/runtime_paths.py"),
