+++ examples/new-project/.loom/bootstrap/init-result.json
@@ -207,6 +207,11 @@
         "owner": "loom-runtime"
       },
       {
+        "path": ".loom/bin/github_host.py",
+        "kind": "loom-tool-support",
+        "owner": "loom-runtime"
+      },
+      {
         "path": ".loom/bin/loom_flow.py",
         "kind": "loom-tool",
         "owner": "loom-runtime"
@@ -426,7 +431,7 @@
       "path": ".loom/bin/loom_init.py",
       "kind": "loom-tool",
       "source": "skills/shared/scripts/loom_init.py",
-      "sha256": "781fd54acbf8a26f9254b6a7a1cfb63933f2a57626116cf3008436da041dfb35"
+      "sha256": "320818f8542430f2348ac6ebd1f65c254f89c193404a13e4caa13058623f383a"
     },
     {
       "path": ".loom/bin/fact_chain_support.py",
@@ -447,10 +452,16 @@
       "sha256": "6ea94ddc6bd7408427c97b4322eb53f646f29f0d2dc3d702edc65bfcd36533ed"
     },
     {
+      "path": ".loom/bin/github_host.py",
+      "kind": "loom-tool-support",
+      "source": "skills/shared/scripts/github_host.py",
+      "sha256": "b2dd0f23450c682f9d4251344145d5b36b58ac35ba6b61a1b152a00a8cf98cba"
+    },
+    {
       "path": ".loom/bin/loom_flow.py",
       "kind": "loom-tool",
       "source": "skills/shared/scripts/loom_flow.py",
-      "sha256": "ab1cf54c27bb958b7bbf03012d5bd39c103091130aa237b471f7c8fe25a22f86"
+      "sha256": "6cf9a86e99f946241052c3577908b7ef1cb81ae603ff9723378c9f46f6446910"
     },
     {
       "path": ".loom/bin/loom_status.py",
@@ -468,13 +479,13 @@
       "path": ".loom/bin/runtime_state.py",
       "kind": "loom-tool-support",
       "source": "skills/shared/scripts/runtime_state.py",
-      "sha256": "4403fe883b2a8296d4099521d6f508a35943f9f89d56f69ca88cdc7aacf2973e"
+      "sha256": "03ea3ed268b5305a93e04529c17a9681323a04e53609309a22e85d32e09532d9"
     },
     {
       "path": ".loom/bin/loom_check.py",
       "kind": "loom-tool",
       "source": "skills/shared/scripts/loom_check.py",
-      "sha256": "1fe4302e243054d6b1416cc139faa201bf5d796953b77f75adaab447030e5320"
+      "sha256": "ad0f71ba440609e3a7b5444abf29e3044c198e99545ead1ea7fd1f9c32382cfa"
     },
     {
       "path": ".loom/bin/loom_story_carriers.py",
@@ -2181,6 +2192,12 @@
       "requires_intent": null
     },
     {
+      "path": ".loom/bin/github_host.py",
+      "kind": "loom-tool-support",
+      "owner": "loom-runtime",
+      "requires_intent": null
+    },
+    {
       "path": ".loom/bin/loom_flow.py",
       "kind": "loom-tool",
       "owner": "loom-runtime",
@@ -2409,6 +2426,11 @@
       "owner": "loom-runtime"
     },
     {
+      "path": ".loom/bin/github_host.py",
+      "kind": "loom-tool-support",
+      "owner": "loom-runtime"
+    },
+    {
       "path": ".loom/bin/loom_flow.py",
       "kind": "loom-tool",
       "owner": "loom-runtime"
@@ -2585,6 +2607,7 @@
             ".loom/bin/fact_chain_support.py",
             ".loom/bin/governance_surface.py",
             ".loom/bin/github_admission.py",
+            ".loom/bin/github_host.py",
             ".loom/bin/loom_flow.py",
             ".loom/bin/loom_status.py",
             ".loom/bin/runtime_paths.py",
@@ -2677,6 +2700,7 @@
       ".loom/bin/fact_chain_support.py",
       ".loom/bin/governance_surface.py",
       ".loom/bin/github_admission.py",
+      ".loom/bin/github_host.py",
       ".loom/bin/loom_flow.py",
       ".loom/bin/loom_status.py",
       ".loom/bin/runtime_paths.py",
@@ -2829,6 +2853,7 @@
           ".loom/bin/fact_chain_support.py",
           ".loom/bin/governance_surface.py",
           ".loom/bin/github_admission.py",
+          ".loom/bin/github_host.py",
           ".loom/bin/loom_flow.py",
           ".loom/bin/loom_status.py",
           ".loom/bin/runtime_paths.py",
