+++ plugins/loom/.codex-plugin/plugin.json
@@ -23,7 +23,7 @@
     "source_git_sha": "unreleased",
     "source_git_sha_status": "pending_release_commit",
     "plugin_payload_version": "0.28.0",
-    "plugin_payload_hash": "90100e6a38b20a40878e66044db880eb3d9985de293bf5a8b9c09102adc84512",
+    "plugin_payload_hash": "5458291459f134299ac51cb5bf169623bd3d1fd1443d116861997a839916709d",
     "repository_payload": false
   },
   "keywords": [
