+++ src/skills/shared/scripts/runtime_state.py
@@ -37,6 +37,7 @@ EXPECTED_SHARED_RUNTIME_SCRIPTS = (
     "fact_chain_support.py",
     "governance_surface.py",
     "github_admission.py",
+    "github_host.py",
     "loom_flow.py",
     "loom_status.py",
     "loom_check.py",
@@ -48,6 +49,7 @@ EXPECTED_BOOTSTRAP_RUNTIME_SOURCES = {
     ".loom/bin/fact_chain_support.py": "skills/shared/scripts/fact_chain_support.py",
     ".loom/bin/governance_surface.py": "skills/shared/scripts/governance_surface.py",
     ".loom/bin/github_admission.py": "skills/shared/scripts/github_admission.py",
+    ".loom/bin/github_host.py": "skills/shared/scripts/github_host.py",
     ".loom/bin/loom_flow.py": "skills/shared/scripts/loom_flow.py",
     ".loom/bin/loom_status.py": "skills/shared/scripts/loom_status.py",
     ".loom/bin/runtime_paths.py": "skills/shared/scripts/runtime_paths.py",
